requirements.txt
Flask
requests
pyjwt
redis
python-decouple
pika
pymysql
pytz